init integrations microservice (#71)
