--- apps/be/project.json
@@ -1,62 +1,0 @@
-{
-  "name": "be",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "apps/be/src",
-  "projectType": "application",
-  "targets": {
-    "build": {
-      "executor": "@nrwl/webpack:webpack",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "target": "node",
-        "compiler": "tsc",
-        "outputPath": "dist/apps/be",
-        "main": "apps/be/src/main.ts",
-        "tsConfig": "apps/be/tsconfig.app.json",
-        "assets": ["apps/be/src/assets"]
-      },
-      "configurations": {
-        "production": {
-          "optimization": true,
-          "extractLicenses": true,
-          "inspect": false,
-          "fileReplacements": [
-            {
-              "replace": "apps/be/src/environments/environment.ts",
-              "with": "apps/be/src/environments/environment.prod.ts"
-            }
-          ]
-        }
-      }
-    },
-    "serve": {
-      "executor": "@nrwl/js:node",
-      "options": {
-        "buildTarget": "be:build"
-      },
-      "configurations": {
-        "production": {
-          "buildTarget": "be:build:production"
-        }
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "outputFile": "reports/lint-results/be.json",
-        "format": "json",
-        "lintFilePatterns": ["apps/be/**/*.ts"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/apps/be"],
-      "options": {
-        "jestConfig": "apps/be/jest.config.ts",
-        "passWithNoTests": true
-      }
-    }
-  },
-  "tags": []
-}
