--- apps/be/src/app/app.controller.ts
@@ -1,13 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-
-import { AppService } from './app.service';
-
-@Controller()
-export class AppController {
-  constructor(private readonly appService: AppService) {}
-
-  @Get()
-  getData() {
-    return this.appService.getData();
-  }
-}
