--- apps/be/src/main.ts
@@ -1,37 +1,0 @@
-/**
- * This is not a production server yet!
- * This is only a minimal backend to get started.
- */
-
-import { Logger } from '@nestjs/common';
-import { NestFactory } from '@nestjs/core';
-import * as cookieParser from 'cookie-parser';
-import { AppModule } from './app/app.module';
-import { AuthGuard } from '@syncit2.0/core/nest-lib';
-import { Transport } from '@nestjs/microservices';
-
-async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
-  app.use(cookieParser());
-  app.useGlobalGuards(new AuthGuard());
-  const globalPrefix = 'api';
-  app.setGlobalPrefix(globalPrefix);
-  const port = process.env.PORT || 3333;
-  app.connectMicroservice({
-    transport: Transport.RMQ,
-    options: {
-      urls: [`amqp://${process.env.RABBITMQ_HOST}`],
-      queue: 'event_change',
-      queueOptions: {
-        durable: false,
-      },
-    },
-  });
-  await app.startAllMicroservices();
-  await app.listen(port);
-  Logger.log(
-    `🚀 Application is running on: http://localhost:${port}/${globalPrefix}`
-  );
-}
-
-bootstrap();
