+++ apps/microservices/integrations/jest.config.ts
@@ -1,7 +1,7 @@
 /* eslint-disable */
 export default {
-  displayName: 'be',
-  preset: '../../jest.preset.js',
+  displayName: 'microservices-integrations',
+  preset: '../../../jest.preset.js',
   globals: {
     'ts-jest': {
       tsconfig: '<rootDir>/tsconfig.spec.json',
@@ -12,5 +12,5 @@ export default {
     '^.+\\.[tj]s$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/apps/be',
+  coverageDirectory: '../../../coverage/apps/microservices/integrations',
 };
