+++ apps/microservices/integrations/src/app/app.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { IntegrationsController } from './app.controller';
+
+@Module({
+  imports: [],
+  controllers: [IntegrationsController],
+})
+export class AppModule {}
